requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
multidict==6.0.4
packaging==23.1
Pillow==10.0.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
pypng==0.20220715.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
sqlparse==0.4.4
swagger-spec-validator==3.0.3
twilio==8.8.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
yarl==1.9.2
